// src/chat/enums/chat-message-format.enum.ts
export enum ChatMessageKind {
  TEXT = 'text',
  CODE = 'code',
  HTML = 'html',
  JSON = 'json',
  MARKDOWN = 'markdown',
}

export enum ChatLanguage {
  PLAINTEXT = 'plaintext',
  JAVASCRIPT = 'javascript',
  TYPESCRIPT = 'typescript',
  JSON = 'json',
  HTML = 'html',
  CSS = 'css',
  SQL = 'sql',
  BASH = 'bash',
  PYTHON = 'python',
  JAVA = 'java',
  CSHARP = 'csharp',
  CPP = 'cpp',
}